requirements.txt
Django==1.5.0
South==0.7.6
psycopg2==2.4.6
django-jenkins==0.14.0
django-debug-toolbar==0.9.4
pep8==1.4.5
pyflakes==0.6.1
django-braces==1.0.0
django-widget_tweaks==1.1.2
PIL==1.1.7
django-iban==0.1.5